[143] switch theme

+++ core-component/persistence/src/main/kotlin/dev/zitech/core/persistence/framework/database/dao/UserAccountDao.kt
@@ -43,5 +43,5 @@ internal interface UserAccountDao {
     suspend fun removeCurrentUserAccount(): Int
 
     @Query("UPDATE user_accounts SET theme=:theme WHERE isCurrentUserAccount = 1")
-    suspend fun updateCurrentUserAccountTheme(theme: Long)
+    suspend fun updateCurrentUserAccountTheme(theme: Int)
 }
