+++ core-component/persistence/src/test/kotlin/dev/zitech/core/persistence/framework/database/dao/FakeUserAccountDao.kt
@@ -49,7 +49,7 @@ internal class FakeUserAccountDao : UserAccountDao {
         }
     }
 
-    override suspend fun updateCurrentUserAccountTheme(theme: Long) {
+    override suspend fun updateCurrentUserAccountTheme(theme: Int) {
         val currentUserAccount = getCurrentUserAccount()
         val updatedCurrentUserAccount = UserAccountEntityFactory.createUserAccountEntity(
             id = currentUserAccount?.id,
