+++ core-component/persistence/src/test/kotlin/dev/zitech/core/persistence/framework/database/entity/UserAccountEntityFactory.kt
@@ -24,7 +24,7 @@ internal object UserAccountEntityFactory {
     fun createUserAccountEntity(
         id: Long? = DataFactory.createRandomLong(),
         isCurrentUserAccount: Boolean = false,
-        theme: Long = 0L
+        theme: Int = 0
     ) = UserAccountEntity(
         id,
         isCurrentUserAccount,
